Revert "split types into multiple files"
This reverts commit e1b8661fd3da0f18e9b465de6d2d3f2d151b8bec.
