--- DiscordWebhooks/Models/WebhookAuthor.cs
@@ -1,10 +1,0 @@
-﻿// namespace required for backward compatibility
-namespace ShimmyMySherbet.DiscordWebhooks.Embeded
-{
-	public class WebhookAuthor
-	{
-		public string name;
-		public string url;
-		public string icon_url;
-	}
-}
