--- DiscordWebhooks/Models/WebhookField.cs
@@ -1,12 +1,0 @@
-﻿// namespace required for backward compatibility
-namespace ShimmyMySherbet.DiscordWebhooks.Embeded
-{
-	public class WebhookField
-	{
-		public string name;
-
-		public string value;
-
-		public bool inline;
-	}
-}
