--- DiscordWebhooks/Models/WebhookFooter.cs
@@ -1,10 +1,0 @@
-﻿// namespace required for backward compatibility
-namespace ShimmyMySherbet.DiscordWebhooks.Embeded
-{
-	public class WebhookFooter
-	{
-		public string text;
-
-		public string icon_url;
-	}
-}
