--- DiscordWebhooks/Models/WebhookImage.cs
@@ -1,9 +1,0 @@
-﻿// namespace required for backward compatibility
-namespace ShimmyMySherbet.DiscordWebhooks.Embeded
-{
-	public class WebhookImage
-	{
-		public string url = "";
-	}
-
-}
